+++ rendering/cases/text-style-overlap/main.js
@@ -19,7 +19,7 @@ const vectorSource = new VectorSource();
 const pointStyle = new Style({
   text: new Text({
     text: 'Point Label',
-    font: '"Ubuntu Regular", sans-serif',
+    font: '"Ubuntu Regular"',
     fill: new Fill({
       color: 'red'
     }),
@@ -32,7 +32,7 @@ const lineStyle = new Style({
   stroke: new Stroke({color: 'blue'}),
   text: new Text({
     text: 'Line Label',
-    font: '"Ubuntu Regular", sans-serif',
+    font: '"Ubuntu Regular"',
     fill: new Fill({
       color: 'red'
     }),
