+++ rendering/cases/text-style/main.js
@@ -20,7 +20,7 @@ feature = new Feature({
 feature.setStyle(new Style({
   text: new Text({
     text: 'hello',
-    font: '14px "Ubuntu Regular", sans-serif',
+    font: '14px "Ubuntu Regular"',
     scale: 2,
     fill: new Fill({
       color: 'red'
@@ -39,7 +39,7 @@ feature = new Feature({
 feature.setStyle(new Style({
   text: new Text({
     text: 'upside down',
-    font: '14px "Ubuntu Regular", sans-serif',
+    font: '14px "Ubuntu Regular"',
     rotation: Math.PI,
     stroke: new Stroke({
       color: 'red',
@@ -55,7 +55,7 @@ feature = new Feature({
 });
 feature.setStyle(new Style({
   text: new Text({
-    font: '"Ubuntu Regular", sans-serif',
+    font: '"Ubuntu Regular"',
     text: 'rotateWithView',
     rotateWithView: true,
     stroke: new Stroke({
@@ -72,7 +72,7 @@ feature = new Feature({
 feature.setStyle(new Style({
   text: new Text({
     text: 'hello',
-    font: '14px "Ubuntu Regular", sans-serif',
+    font: '14px "Ubuntu Regular"',
     padding: [1, 2, 3, 5],
     backgroundFill: new Fill({
       color: 'rgba(55, 55, 55, 0.25)'
@@ -92,7 +92,7 @@ feature = new Feature({
 feature.setStyle(new Style({
   text: new Text({
     text: 'hello',
-    font: '14px "Ubuntu Regular", sans-serif',
+    font: '14px "Ubuntu Regular"',
     padding: [1, 2, 3, 5],
     backgroundFill: new Fill({
       color: 'rgba(55, 55, 55, 0.25)'
