+++ rendering/default/index.html
@@ -20,7 +20,6 @@
         font-family: "Ubuntu Regular";
         src: url("/data/fonts/ubuntu-regular-webfont.eot");
         src: url("/data/fonts/ubuntu-regular-webfont.woff") format("woff"),
-          url("/data/fonts/ubuntu-regular-webfont.otf") format("opentype"),
           url("/data/fonts/ubuntu-regular-webfont.svg#filename") format("svg");
       }
       @font-face {
@@ -29,7 +28,6 @@
         font-style: italic;
         src: url("/data/fonts/ubuntu-bold-italic-webfont.eot");
         src: url("/data/fonts/ubuntu-bold-italic-webfont.woff") format("woff"),
-          url("/data/fonts/ubuntu-bold-italic-webfont.otf") format("opentype"),
           url("/data/fonts/ubuntu-bold-italic-webfont.svg#filename") format("svg");
       }
     </style>
